% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_sightings.R
\name{agg.sightings}
\alias{agg.sightings}
\title{aggregate sightings}
\usage{
agg.sightings(surveys, focals)
}
\arguments{
\item{surveys}{db download of the survey dolphin file from MMDD}

\item{focals}{db download of the focal dolphin file from MMDD}
}
\description{
aggregate sightings
}
