% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sighting_check.R
\name{sighting.check}
\alias{sighting.check}
\title{sighting check}
\usage{
sighting.check(sightings, demography)
}
\arguments{
\item{sightings}{df of sighitngs with Observation.Date and Dolphin.ID as columns - output from sightings function}

\item{demography}{db download of life history file from MMDD}

\item{focals}{db download of the focal dolphin file from MMDD}
}
\description{
sighting check
}
