% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_died_function.R
\name{who.died}
\alias{who.died}
\title{who died}
\usage{
who.died(surveys, focals, demography, time.cutoff)
}
\arguments{
\item{surveys}{db download of the survey dolphin file from MMDD}

\item{focals}{db download of the focal dolphin file from MMDD}

\item{demography}{db download of life history file from MMDD}

\item{time.cutoff}{date we want to use to calculate time since the last sighting}
}
\description{
who died
}
