% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years_sighted.R
\name{years.sighted}
\alias{years.sighted}
\title{years sighted}
\usage{
years.sighted(surveys, focals)
}
\arguments{
\item{surveys}{db download of the survey dolphin file from MMDD}

\item{focals}{db download of the focal dolphin file from MMDD}
}
\description{
years sighted
}
