% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sighting_check.R
\name{sighting.check}
\alias{sighting.check}
\title{sighting check}
\usage{
sighting.check(sightings, demography)
}
\arguments{
\item{sightings}{df of sightings with Observation.Date and Dolphin.ID as columns - output from sightings function}

\item{demography}{db download of life history file from MMDD}
}
\description{
outputs a df of observations that occurred before an individual's birth or after an individual's death in order to catch date typos or resurrections.
}
