% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sighting_check.R
\name{sighting.check}
\alias{sighting.check}
\title{sighting check}
\usage{
sighting.check(surveys, focals, demography)
}
\arguments{
\item{surveys}{db download of the survey dolphin file from MMDD}

\item{focals}{db download of the focal dolphin file from MMDD}

\item{demography}{db download of life history file from MMDD}
}
\description{
sighting check
}
