% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sightings.R
\name{sightings}
\alias{sightings}
\title{create sighting df}
\usage{
sightings(surveys, focals)
}
\arguments{
\item{surveys}{db download of the survey dolphin file from MMDD}

\item{focals}{db download of the focal dolphin file from MMDD}
}
\description{
outputs a dataframe containing each individuals sighting history
}
