% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years_sighted.R
\name{years.sighted}
\alias{years.sighted}
\title{years sighted}
\usage{
years.sighted(sightings)
}
\arguments{
\item{sightings}{df of sighitngs with Observation.Date and Dolphin.ID as columns - output from sightings function}
}
\description{
years sighted
}
