% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years_sighted.R
\name{years.sighted}
\alias{years.sighted}
\title{years sighted}
\usage{
years.sighted(sightings)
}
\arguments{
\item{sightings}{df of sightings with Observation.Date and Dolphin.ID as columns - output from sightings function}
}
\description{
Outputs a df of all of the years each individual was sighted, and the gaps between those years.
Used for assigning death dates.
}
